+++ scilpy/tractanalysis/scoring.py
@@ -1,23 +1,13 @@
-import logging
+# -*- coding: utf-8 -*-
 
-import nibabel as nib
+import logging
 import numpy as np
-import os
 
 from dipy.io.stateful_tractogram import StatefulTractogram
-from dipy.tracking.utils import length
-from scipy.ndimage import binary_dilation
-from sklearn.cluster import KMeans
 
-from scilpy.io.image import get_data_as_mask
-from scilpy.io.streamlines import load_tractogram_with_reference
-from scilpy.segment.streamlines import filter_grid_roi, filter_grid_roi_both
-from scilpy.tracking.tools import filter_streamlines_by_total_length_per_dim
-from scilpy.tractanalysis.features import remove_loops_and_sharp_turns
 from scilpy.tractanalysis.reproducibility_measures import \
     get_endpoints_density_map
 from scilpy.tractanalysis.streamlines_metrics import compute_tract_counts_map
-from scilpy.utils.filenames import split_name_with_nii
 
 
 def compute_f1_score(overlap, overreach):
@@ -167,409 +157,150 @@ def get_binary_maps(streamlines, sft):
     return bundles_voxels, endpoints_voxels
 
 
-def compute_masks(gt_files, parser, args):
+def compute_tractometry(all_vs_ids, all_wpc_ids, all_ic_ids, all_nc_ids,
+                        vs_ids_list, wpc_ids_list, ic_ids_list,
+                        vb_sft_list, wpc_sft_list, ib_sft_list, sft, args,
+                        bundles_names, gt_masks, dimensions, comb_filename):
+
     """
-    Compute ground-truth masks. If the file is already a mask, load it.
-    If it is a bundle, compute the mask. If the filename is None, appends None
-    to the lists of masks.
-
-    Parameters
-    ----------
-    gt_files: list
-        List of either StatefulTractograms or niftis.
-    parser: ArgumentParser
-        Argument parser which handles the script's arguments.
-    args: Namespace
-        List of arguments passed to the script.
-
-    Returns
-    -------
-    mask: numpy.ndarray
-        The loaded mask.
-    inv_mask: numpy.ndarray
-        The inverted mask (contains True where mask was False and vice-versa).
-    affine: the affine of the last image.
-    dimensions: the dimensions of the last image.
+    Tractometry stats: First in terms of connections (NC, IC, VS, WPC), then
+    in terms of volume (OL, OR, Dice score)
     """
-    save_ref = args.reference
+    logging.info("Computing tractometry")
 
-    gt_bundle_masks = []
-    gt_bundle_inv_masks = []
+    nb_bundles = len(bundles_names)
 
-    affine = None
-    dimensions = None
-    for gt_bundle in gt_files:
-        if gt_bundle is not None:
-            # Support ground truth as streamlines or masks
-            # Will be converted to binary masks immediately
-            _, ext = split_name_with_nii(gt_bundle)
-            if ext in ['.gz', '.nii.gz']:
-                gt_img = nib.load(gt_bundle)
-                gt_mask = get_data_as_mask(gt_img)
+    # Total number of streamlines for each category
+    vs_count = len(all_vs_ids)
+    wpc_count = len(all_wpc_ids)
+    ic_count = len(all_ic_ids)
+    nc_count = len(all_nc_ids)
+    total_count = len(sft)
 
-                if affine is not None:
-                    # compare affines.
-                    # todO
-                    logging.debug('Previous affine discarded. (todo)')
-                affine = gt_img.affine
-                dimensions = gt_mask.shape
-            else:
-                # Cheating ref because it may send a lot of warning if loading
-                # many trk with ref (reference was maybe added only for some
-                # of these files)
-                if ext == '.trk':
-                    args.reference = None
+    final_results = {
+        "tractogram_filename": str(args.in_tractogram),
+        "total_streamlines": total_count,
+        "VB": len([x for x in vs_ids_list if len(x) > 0]),
+        "VS": vs_count,
+        "VS_ratio": vs_count / total_count,
+        "IS": ic_count + nc_count,  # ic_count = 0 if not args.compute_ic
+        "IS_ratio": (ic_count + nc_count) / total_count,
+    }
+
+    if args.compute_ic:
+        final_results.update({
+            "IB": len([x for x in ic_ids_list if len(x) > 0]),
+            "IC": ic_count,
+            "IC_ratio": ic_count / total_count,
+            "NC": nc_count,
+            "NC_ratio": nc_count / total_count})
+
+    if args.save_wpc_separately:
+        final_results.update({
+            "WPC": wpc_count,
+            "WPC_bundle": len([x for x in wpc_ids_list if len(x) > 0]),
+            "WPC_ratio": wpc_count / total_count})
+
+    # Tractometry stats over volume: OL, OR, Dice score
+    mean_overlap = 0.0
+    mean_overreach_gt = 0.0
+    mean_overreach_n = 0.0
+    mean_f1 = 0.0
+
+    bundle_wise_dict = {}
+    for i in range(nb_bundles):
+        current_vb = vb_sft_list[i].streamlines
+        bundle_results = {"VS": len(current_vb)}
+        if gt_masks[i] is not None:
+            # Getting the recovered mask
+            current_vb_voxels, current_vb_endpoints_voxels = get_binary_maps(
+                current_vb, sft)
+
+            (f1, tp_nb_voxels, fp_nb_voxels, fn_nb_voxels,
+             overlap, overreach_pct_gt, overreach_pct_total) = \
+                compute_f1_overlap_overreach(
+                    current_vb_voxels, gt_masks[i], dimensions)
+
+            # Endpoints coverage
+            # todo. What is this? Useful?
+            endpoints_overlap = gt_masks[i] * current_vb_endpoints_voxels
+            endpoints_overreach = np.zeros(dimensions)
+            endpoints_overreach[np.where(
+                (gt_masks[i] == 0) & (current_vb_endpoints_voxels >= 1))] = 1
+
+            bundle_results.update({
+                "TP": tp_nb_voxels,
+                "FP": fp_nb_voxels,
+                "FN": fn_nb_voxels,
+                "OL": overlap,
+                "OR_gt": overreach_pct_gt,
+                "ORn": overreach_pct_total,
+                "f1": f1,
+                "endpoints_OL": np.count_nonzero(endpoints_overlap),
+                "endpoints_OR": np.count_nonzero(endpoints_overreach)
+            })
+
+            # WPC
+            if args.save_wpc_separately:
+                wpc = wpc_sft_list[i]
+                if wpc is not None and len(wpc.streamlines) > 0:
+                    current_wpc_streamlines = wpc.streamlines
+                    current_wpc_voxels, _ = get_binary_maps(
+                        current_wpc_streamlines, sft)
+
+                    # We could add an option to include wpc streamlines to the
+                    # overreach count. But it seems more natural to exclude wpc
+                    # streamlines from any count. Separating into a different
+                    # statistic dict.
+                    (_, tp_nb_voxels, fp_nb_voxels, _, overlap,
+                     overreach_pct_gt, overreach_pct_total) = \
+                        compute_f1_overlap_overreach(
+                            current_vb_voxels, gt_masks[i], dimensions)
+
+                    wpc_results = {
+                        "Count": len(current_wpc_streamlines),
+                        "TP": tp_nb_voxels,
+                        "FP": fp_nb_voxels,
+                        "OL": overlap,
+                        "OR_gt": overreach_pct_gt,
+                        "ORn": overreach_pct_total,
+                    }
+                    bundle_results.update({"WPC": wpc_results})
                 else:
-                    args.reference = save_ref
-                gt_sft = load_tractogram_with_reference(
-                    parser, args, gt_bundle, bbox_check=False)
-                gt_sft.to_vox()
-                gt_sft.to_corner()
-                _affine, _dimensions, _, _ = gt_sft.space_attributes
-                if affine is not None:
-                    # compare affines.
-                    # todO
-                    logging.debug('Previous affine discarded. (todo)')
-                affine = _affine
-                dimensions = _dimensions
-                gt_mask = compute_tract_counts_map(gt_sft.streamlines,
-                                                   dimensions).astype(np.int16)
-            gt_inv_mask = np.zeros(dimensions, dtype=np.int16)
-            gt_inv_mask[gt_mask == 0] = 1
-            gt_mask[gt_mask > 0] = 1
-        else:
-            gt_mask = None
-            gt_inv_mask = None
-
-        gt_bundle_masks.append(gt_mask)
-        gt_bundle_inv_masks.append(gt_inv_mask)
-
-    return gt_bundle_masks, gt_bundle_inv_masks, affine, dimensions
-
-
-def split_heads_tails_kmeans(data):
-    """
-    Split a mask between head and tail with k means.
-
-    Parameters
-    ----------
-    data: numpy.ndarray
-        Mask to be split.
-
-    Returns
-    -------
-    mask_1: numpy.ndarray
-        "Head" of the mask.
-    mask_2: numpy.ndarray
-        "Tail" of the mask.
-    """
-
-    X = np.argwhere(data)
-    k_means = KMeans(n_clusters=2).fit(X)
-    mask_1 = np.zeros(data.shape)
-    mask_2 = np.zeros(data.shape)
-
-    mask_1[tuple(X[np.where(k_means.labels_ == 0)].T)] = 1
-    mask_2[tuple(X[np.where(k_means.labels_ == 1)].T)] = 1
-
-    return mask_1, mask_2
-
-
-def extract_tails_heads_from_endpoints(gt_endpoints, out_dir):
-    """
-    Extract two masks from a single mask containing two regions.
-
-    Parameters
-    ----------
-    gt_endpoints: str
-        Ground-truth mask filename.
-    out_dir: str
-        Path where to save the heads and tails.
-
-    Returns
-    -------
-    tails: list
-        List of tail filenames.
-    heads: list
-        List of head filenames.
-    affine: numpy.ndarray
-        Affine of mask image.
-    dimensions: tuple of int
-        Dimensions of the mask image.
-    """
-    mask_img = nib.load(gt_endpoints)
-    mask = get_data_as_mask(mask_img)
-    affine = mask_img.affine
-    dimensions = mask.shape
-
-    head, tail = split_heads_tails_kmeans(mask)
-
-    basename = os.path.basename(
-        split_name_with_nii(gt_endpoints)[0])
-    tail_filename = os.path.join(
-        out_dir, '{}_tail.nii.gz'.format(basename))
-    head_filename = os.path.join(
-        out_dir, '{}_head.nii.gz'.format(basename))
-    nib.save(nib.Nifti1Image(head.astype(
-        mask.dtype), affine), head_filename)
-    nib.save(nib.Nifti1Image(tail.astype(
-        mask.dtype), affine), tail_filename)
-
-    return tail_filename, head_filename, affine, dimensions
-
-
-def compute_endpoint_masks(roi_options, affine, dimensions, out_dir):
-    """
-    If endpoints without heads/tails are loaded, split them and continue
-    normally after. Q/C of the output is important.
-
-    Parameters
-    ------
-    roi_options: dict
-        Keys are the bundle names. For each bundle, the value is itself a
-        dictionary either key 'gt_endpoints' (the name of the file
-        containing the bundle's endpoints), or both keys 'gt_tail' and
-        'gt_head' (the names of the respetive files).
-    affine: array
-        A nibabel affine. Final masks must be compatible.
-    dimensions: array
-        A nibabel dimensions. Final masks must be compatible.
-    out_dir: str
-        Where to save the heads and tails.
-
-    Returns
-    -------
-    tails, heads: lists of filenames with length the number of bundles.
-    """
-    tails = []
-    heads = []
-    for bundle_options in roi_options:
-        if 'gt_endpoints' in bundle_options:
-            tail, head, _affine, _dimensions = \
-                extract_tails_heads_from_endpoints(
-                    bundle_options['gt_endpoints'], out_dir)
-            if affine is not None:
-                # Compare affine
-                # todo
-                pass
-            logging.debug('_affine discarded. (todo)')
-            if dimensions is not None:
-                # Compare dimensions
-                # todo
-                pass
-            logging.debug("_dimensions discarded (todo)")
-        else:
-            tail = bundle_options['gt_tail']
-            head = bundle_options['gt_head']
-
-        tails.append(tail)
-        heads.append(head)
-
-    return tails, heads
-
-
-def extract_vb_vs_one_bundle(
-        sft, head_filename, tail_filename, limits_length, angle,
-        orientation_length, abs_orientation_length, inv_all_mask,
-        any_mask, dilate_endpoints):
-    """
-    Extract valid bundle (and valid streamline ids) from a tractogram, based
-    on two regions of interest for the endpoints, one region of interest for
-    the inclusion of streamlines, and maximum length, maximum angle,
-    maximum length per orientation.
-
-    Parameters
-    ----------
-    sft: StatefulTractogram
-        Tractogram containing the streamlines to be extracted.
-    head_filename: str
-        Filename of the "head" of the bundle.
-    tail_filename: str
-        Filename of the "tail" of the bundle.
-    limits_length: list or None
-        Bundle's length parameters: [min max].
-    angle: int or None
-        Bundle's max angle.
-    orientation_length: list or None
-        Bundle's length parameters in each direction:
-        [[min_x, max_x], [min_y, max_y], [min_z, max_z]]
-    abs_orientation_length: idem, computed in absolute values.
-    inv_all_mask: np.ndarray or None
-        Inverse ALL mask for this bundle: no point must be outside the mask.
-    any_mask: np.ndarray or None
-        ANY mask for this bundle.
-        Streamlines must pass through this mask (touch it) to be included
-        in the bundle.
-    dilate_endpoints: int or None
-        If set, dilate the masks for n iterations.
-
-    Returns
-    -------
-    vs_ids: list
-        List of ids of valid streamlines
-    wpc_ids: list
-        List of ids of wrong-path connections
-    bundle_stats: dict
-        Dictionary of recognized streamlines statistics
-    """
-    mask_1_img = nib.load(head_filename)
-    mask_2_img = nib.load(tail_filename)
-    mask_1 = get_data_as_mask(mask_1_img)
-    mask_2 = get_data_as_mask(mask_2_img)
-
-    if dilate_endpoints:
-        mask_1 = binary_dilation(mask_1, iterations=dilate_endpoints)
-        mask_2 = binary_dilation(mask_2, iterations=dilate_endpoints)
-
-    _, vs_ids = filter_grid_roi_both(sft, mask_1, mask_2)
-
-    wpc_ids = []
-    bundle_stats = {"Initial count head to tail": len(vs_ids)}
-
-    # Remove out of inclusion mask (limits_mask)
-    if len(vs_ids) > 0 and inv_all_mask is not None:
-        tmp_sft = sft[vs_ids]
-        _, out_of_mask_ids_from_vs = filter_grid_roi(
-            tmp_sft, inv_all_mask, 'any', False)
-        out_of_mask_ids = vs_ids[out_of_mask_ids_from_vs]
-
-        bundle_stats.update({"WPC_out_of_mask": len(out_of_mask_ids)})
-
-        # Update ids
-        wpc_ids.extend(out_of_mask_ids)
-        vs_ids = np.setdiff1d(vs_ids, wpc_ids)
-
-    # Remove streamlines not passing through any_mask
-    if len(vs_ids) > 0 and any_mask is not None:
-        tmp_sft = sft[vs_ids]
-        _, in_mask_ids_from_vs = filter_grid_roi(
-            tmp_sft, any_mask, 'any', False)
-        in_mask_ids = vs_ids[in_mask_ids_from_vs]
-
-        out_of_mask_ids = np.setdiff1d(vs_ids, in_mask_ids)
-        bundle_stats.update({"WPC_not_reaching_the_ANY_mask":
-                                 len(out_of_mask_ids)})
-
-        # Update ids
-        wpc_ids.extend(out_of_mask_ids)
-        vs_ids = in_mask_ids
-
-    # Remove invalid lengths
-    if len(vs_ids) > 0 and limits_length is not None:
-        min_len, max_len = limits_length
-
-        # Bring streamlines to world coordinates so proper length
-        # is calculated
-        sft.to_rasmm()
-        lengths = np.array(list(length(sft.streamlines[vs_ids])))
-        sft.to_vox()
-
-        # Compute valid lengths
-        valid_length_ids_mask_from_vs = np.logical_and(lengths > min_len,
-                                                       lengths < max_len)
-
-        bundle_stats.update({
-            "WPC_invalid_length": int(sum(~valid_length_ids_mask_from_vs))})
-
-        # Update ids
-        wpc_ids.extend(vs_ids[~valid_length_ids_mask_from_vs])
-        vs_ids = vs_ids[valid_length_ids_mask_from_vs]
-
-    # Remove invalid lengths per orientation
-    if len(vs_ids) > 0 and orientation_length is not None:
-        # Compute valid lengths
-        limits_x, limits_y, limits_z = orientation_length
-
-        _, valid_orientation_ids_from_vs, _ = \
-            filter_streamlines_by_total_length_per_dim(
-                sft[vs_ids], limits_x, limits_y, limits_z,
-                use_abs=False, save_rejected=False)
-
-        # Update ids
-        valid_orientation_ids = vs_ids[valid_orientation_ids_from_vs]
-        invalid_orientation_ids = np.setdiff1d(vs_ids, valid_orientation_ids)
-
-        bundle_stats.update({
-            "WPC_invalid_orientation": len(invalid_orientation_ids)})
-
-        wpc_ids.extend(invalid_orientation_ids)
-        vs_ids = valid_orientation_ids
-
-    # Idem in abs
-    if len(vs_ids) > 0 and abs_orientation_length is not None:
-        # Compute valid lengths
-        limits_x, limits_y, limits_z = abs_orientation_length
-
-        _, valid_orientation_ids_from_vs, _ = \
-            filter_streamlines_by_total_length_per_dim(
-                sft[vs_ids], limits_x, limits_y, limits_z,
-                use_abs=True, save_rejected=False)
-
-        # Update ids
-        valid_orientation_ids = vs_ids[valid_orientation_ids_from_vs]
-        invalid_orientation_ids = np.setdiff1d(vs_ids,
-                                               valid_orientation_ids)
-
-        bundle_stats.update({
-            "WPC_invalid_orientation_abs": len(invalid_orientation_ids)})
-
-        wpc_ids.extend(invalid_orientation_ids)
-        vs_ids = valid_orientation_ids
-
-    # Remove loops from tc
-    if len(vs_ids) > 0 and angle is not None:
-        # Compute valid angles
-        valid_angle_ids_from_vs = remove_loops_and_sharp_turns(
-            sft.streamlines[vs_ids], angle)
-
-        # Update ids
-        valid_angle_ids = vs_ids[valid_angle_ids_from_vs]
-        invalid_angle_ids = np.setdiff1d(vs_ids, valid_angle_ids)
-
-        bundle_stats.update({"WPC_invalid_length": len(invalid_angle_ids)})
-
-        wpc_ids.extend(invalid_angle_ids)
-        vs_ids = valid_angle_ids
-
-    bundle_stats.update({"VS": len(vs_ids)})
-
-    return list(vs_ids), list(wpc_ids), bundle_stats
-
-
-def extract_false_connections(sft, mask_1_filename, mask_2_filename,
-                              dilate_endpoints):
-    """
-    Extract false connections based on two regions from a tractogram.
-
-    Parameters
-    ----------
-    sft: StatefulTractogram
-        Tractogram containing the streamlines to be extracted.
-    mask_1_filename: str
-        Filename of the "head" of the bundle.
-    mask_2_filename: str
-        Filename of the "tail" of the bundle.
-    dilate_endpoints: int or None
-        If set, dilate the masks for n iterations.
-
-    Returns
-    -------
-    fc_sft: StatefulTractogram
-        SFT of false connections.
-    sft: StatefulTractogram
-        SFT of remaining streamlines.
-    """
-
-    mask_1_img = nib.load(mask_1_filename)
-    mask_2_img = nib.load(mask_2_filename)
-    mask_1 = get_data_as_mask(mask_1_img)
-    mask_2 = get_data_as_mask(mask_2_img)
-
-    if dilate_endpoints:
-        mask_1 = binary_dilation(mask_1, iterations=dilate_endpoints)
-        mask_2 = binary_dilation(mask_2, iterations=dilate_endpoints)
-
-    _, fc_ids = filter_grid_roi_both(sft, mask_1, mask_2)
-
-    fc_sft = sft[fc_ids]
-    return fc_sft, fc_ids
+                    bundle_results.update({"WPC": None})
+
+        mean_overlap += bundle_results["OL"]
+        mean_overreach_gt += bundle_results["OR_gt"]
+        mean_overreach_n += bundle_results["ORn"]
+        mean_f1 += bundle_results["f1"]
+        bundle_wise_dict.update({bundles_names[i]: bundle_results})
+
+    if args.compute_ic:
+        # -----------
+        # False connections stats: number of voxels
+        # -----------
+        ic_results = {}
+        for i, filename in enumerate(comb_filename):
+            current_ib = ib_sft_list[i].streamlines
+
+            if len(current_ib):
+                current_ib_voxels, _ = get_binary_maps(current_ib, sft)
+
+                bundle_results = {
+                    "filename": filename,
+                    "IC": len(current_ib),
+                    "nb_voxels": np.count_nonzero(current_ib_voxels)
+                }
+                ic_results.update({str(filename): bundle_results})
+
+        bundle_wise_dict.update({"IB": ic_results})
+
+    final_results.update({
+        "bundle_wise": bundle_wise_dict,
+        "mean_OL": mean_overlap / nb_bundles,
+        "mean_OR_gt": mean_overreach_gt / nb_bundles,
+        "mean_ORn": mean_overreach_n / nb_bundles,
+        "mean_f1": mean_f1 / nb_bundles
+    })
+
+    return final_results
